requirements.txt
easydict==1.10
tokenizers==0.12.1
numpy>=1.19.2
ftfy==6.1.1
transformers==4.18.0
imageio==2.15.0
fairscale==0.4.6
ipdb
open-clip-torch==2.0.2
xformers==0.0.13
chardet==5.1.0
torchdiffeq==0.2.3
opencv-python==4.4.0.46
opencv-python-headless==4.7.0.68
torchsde==0.2.6
simplejson==3.18.4
motion-vector-extractor==1.0.6
scikit-learn
scikit-image
rotary-embedding-torch==0.2.1
pynvml==11.5.0
triton==2.0.0.dev20221120
pytorch-lightning==1.4.2
torchmetrics==0.6.0
gradio==3.39.0
imageio-ffmpeg
kornia
tyro
dearpygui
einops
lpips
matplotlib
packaging
Pillow
pygltflib
rembg[gpu,cli]
rich
safetensors
scipy
tqdm
trimesh
kiui >= 0.2.3
roma
plyfile